refactor: MainActivity 에 있던 Composable 함수들 파일로 분리
