+++ app/src/main/java/nextstep/signup/signup/SignUpTextField.kt
@@ -1,0 +1,81 @@
+package nextstep.signup.signup
+
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextField
+import androidx.compose.material3.TextFieldDefaults
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.ui.text.input.PasswordVisualTransformation
+import androidx.compose.ui.text.input.VisualTransformation
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import nextstep.signup.SignUpTextFieldType
+import nextstep.signup.ui.theme.Blue20
+import nextstep.signup.ui.theme.Blue50
+import nextstep.signup.ui.theme.Gray50
+import nextstep.signup.ui.theme.Gray70
+import nextstep.signup.ui.theme.Red50
+
+@Composable
+fun SignUpTextField(
+    text: String,
+    type: SignUpTextFieldType,
+    modifier: Modifier = Modifier,
+    onTextChanged: (String) -> Unit = {},
+    errorMessage: String
+) {
+    TextField(
+        value = text,
+        onValueChange = onTextChanged,
+        label = {
+            Text(
+                text = type.hint,
+                fontSize = 16.sp,
+                lineHeight = 24.sp,
+                fontWeight = FontWeight.W400,
+                letterSpacing = 0.5.sp
+            )
+        },
+        isError = errorMessage.isNotEmpty(),
+        singleLine = true,
+        colors = TextFieldDefaults.colors(
+            focusedTextColor = Gray70,
+            unfocusedTextColor = Gray70,
+            focusedLabelColor = Blue50,
+            unfocusedLabelColor = Gray50,
+            errorLabelColor = Red50,
+            focusedIndicatorColor = Blue50,
+            unfocusedIndicatorColor = Gray50,
+            errorIndicatorColor = Red50,
+            cursorColor = Blue50,
+            errorCursorColor = Red50,
+            focusedContainerColor = Blue20,
+            unfocusedContainerColor = Blue20,
+        ),
+        supportingText = {
+            if (errorMessage.isNotEmpty()) {
+                SignUpTextFieldSupportingText(errorMessage)
+            }
+        },
+        visualTransformation = when (type) {
+            SignUpTextFieldType.USERNAME, SignUpTextFieldType.EMAIL -> VisualTransformation.None
+            SignUpTextFieldType.PASSWORD, SignUpTextFieldType.PASSWORD_CONFIRM -> PasswordVisualTransformation()
+        },
+        keyboardOptions = KeyboardOptions(
+            keyboardType = when (type) {
+                SignUpTextFieldType.USERNAME -> KeyboardType.Text
+                SignUpTextFieldType.EMAIL -> KeyboardType.Email
+                SignUpTextFieldType.PASSWORD, SignUpTextFieldType.PASSWORD_CONFIRM -> KeyboardType.Password
+            }
+        ),
+        modifier = modifier
+            .width(296.dp)
+            .clip(shape = RoundedCornerShape(topStart = 4.dp, topEnd = 4.dp))
+    )
+}
