+++ app/src/main/java/nextstep/signup/signup/SignUpTextFieldSupportingText.kt
@@ -1,0 +1,18 @@
+package nextstep.signup.signup
+
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.sp
+
+@Composable
+fun SignUpTextFieldSupportingText(text: String, modifier: Modifier = Modifier) {
+    Text(
+        text = text,
+        fontSize = 12.sp,
+        lineHeight = 16.sp,
+        fontWeight = FontWeight.W400,
+        modifier = modifier
+    )
+}
