+++ app/src/main/java/nextstep/signup/signup/SignUpTitle.kt
@@ -1,0 +1,22 @@
+package nextstep.signup.signup
+
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.sp
+import nextstep.signup.R
+
+@Composable
+fun SignUpTitle(modifier: Modifier = Modifier) {
+    Text(
+        text = stringResource(R.string.welcome_to_compose),
+        fontSize = 26.sp,
+        fontWeight = FontWeight.W700,
+        color = Color.Black,
+        lineHeight = 20.sp,
+        modifier = modifier,
+    )
+}
